Replace objects structure in DBProvider
